cleaning up bingo logic, react context instead of local storage
