+++ src/index.js
@@ -3,14 +3,15 @@ import ReactDOM from 'react-dom/client';
 import './index.css';
 import App from './App';
 import { SupabaseProvider } from './SupabaseContext';
-
-
+import { BingoProvider } from './BingoContext';
 
 const root = ReactDOM.createRoot(document.getElementById('root'));
 root.render(
   <React.StrictMode>
     <SupabaseProvider>
-      <App />
+      <BingoProvider>
+        <App />
+      </BingoProvider>
     </SupabaseProvider>
   </React.StrictMode>
 );
